feat(mastodon): unified classes
